feat: :sparkles: Added io library
Added dedicated io library
