+++ src/base/exit_code.h
@@ -1,0 +1,13 @@
+#ifndef CHIP8_BASE_H_
+#define CHIP8_BASE_H_
+
+typedef enum {
+    EXIT_CODE_OK = 0,
+    EXIT_CODE_COMMAND_LINE_USAGE_ERROR = 64,
+    EXIT_CODE_ASSEMBLER_ERROR = 65,
+    EXIT_CODE_RUNTIME_ERROR = 70,
+    EXIT_CODE_SYSTEM_ERROR = 71,
+    EXIT_CODE_INPUT_OUTPUT_ERROR = 74
+} chip8_exit_code;
+
+#endif
