--- src/chip8_config.h.in
@@ -1,25 +1,0 @@
-/****************************************************************************
- * Copyright (C) 2023 by Frederik Tobner                                    *
- *                                                                          *
- * This file is part of CHIP-8.                                             *
- *                                                                          *
- * Permission to use, copy, modify, and distribute this software and its    *
- * documentation under the terms of the GNU General Public License is       *
- * hereby granted.                                                          *
- * No representations are made about the suitability of this software for   *
- * any purpose.                                                             *
- * It is provided "as is" without express or implied warranty.              *
- * See the <https://www.gnu.org/licenses/gpl-3.0.html/>GNU General Public   *
- * License for more details.                                                *
- ****************************************************************************/
-
-#ifndef CHIP8_CONFIGURATION_H_
-#define CHIP8_CONFIGURATION_H_
-
-// the configured options and settings for the CHIP-8 emulator
-#define PROJECT_VERSION_MAJOR (@CHIP-8_VERSION_MAJOR@)
-#define PROJECT_VERSION_MINOR (@CHIP-8_VERSION_MINOR@)
-#define PROJECT_VERSION_PATCH (@CHIP-8_VERSION_PATCH@)
-#define PROJECT_NAME ("@PROJECT_NAME@")
-
-#endif
