+++ src/frontend/CMakeLists.txt
@@ -15,12 +15,14 @@ set(FRONTEND_HEADER_FILES
 "table.h"
 )
 
-add_library(${PROJECT_NAME}Frontend STATIC ${FRONTEND_SOURCE_FILES} ${FRONTEND_HEADER_FILES})
+add_library(${PROJECT_NAME}_Frontend STATIC ${FRONTEND_SOURCE_FILES} ${FRONTEND_HEADER_FILES})
+
+target_link_libraries(${PROJECT_NAME}_Frontend ${PROJECT_NAME}_Base)
 
 # Precompiles common.h to speed up compilation of the target
 if(MSVC)
     # VisualStudio only accepts header files that also have a source file    
-    target_precompile_headers(${PROJECT_NAME}Frontend PUBLIC pre_compiled_header.h pre_compiled_header.c) 
+    target_precompile_headers(${PROJECT_NAME}_Frontend PUBLIC pre_compiled_header.h pre_compiled_header.c) 
 else()
-    target_precompile_headers(${PROJECT_NAME}Frontend PUBLIC pre_compiled_header.h)
+    target_precompile_headers(${PROJECT_NAME}_Frontend PUBLIC pre_compiled_header.h)
 endif()
